backend/pyproject.toml
[project]
name = "capstoneproject"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]

requires-python = ">=3.11"
dependencies = ["fastapi", "uvicorn[standard]", "python-dotenv (>=1.1.1,<2.0.0)", "sqlalchemy (>=2.0.43,<3.0.0)", "asyncpg (>=0.30.0,<0.31.0)", "psycopg[binary] (>=3.2.10,<4.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.packages]]
include = "app"
from = "src"